muir-glacier refactored
